docs(template-syntax): drive jade from sample
