+++ public/docs/_examples/template-syntax/ts/package.json
@@ -11,11 +11,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "angular2": "2.0.0-alpha.46",
-    "systemjs": "0.19.2"
+    "angular2": "2.0.0-alpha.48",
+    "systemjs": "0.19.6"
   },
   "devDependencies": {
     "live-server": "^0.8.1",
-    "typescript": "^1.6.2"
+    "typescript": "^1.7.2"
   }
 }
