+++ public/docs/_examples/template-syntax/ts/src/app/boot.ts
@@ -1,0 +1,3 @@
+import {bootstrap} from 'angular2/platform/browser';
+import {AppComponent} from './app.component';
+bootstrap(AppComponent);
