+++ public/docs/_examples/template-syntax/ts/src/index.html
@@ -2,21 +2,19 @@
 <html>
   <head>
     <title>Template Syntax</title>
-    <link rel="stylesheet" href="styles.css">    
+    <link rel="stylesheet" href="styles.css">
     <script src="../node_modules/systemjs/dist/system.src.js"></script>
     <script src="../node_modules/angular2/bundles/angular2.dev.js"></script>
     <script>
       System.config({
         packages: {'app': {defaultExtension: 'js'}}
       });
-      System.import('app/app');
+      System.import('app/boot');
     </script>
   </head>
 
   <body>
     <my-app>Loading...</my-app>
-    <hr>
-    <little-tour>Loading...</little-tour>
   </body>
 
 </html>
