+++ public/docs/_examples/template-syntax/ts/src/styles.css
@@ -4,6 +4,8 @@ img {height: 100px;}
 .child-div {margin-left: 1em; font-weight: normal}
 .hidden {display: none}
 .parent-div {margin-top: 1em; font-weight: bold}
-.special {font-weight:bold;}
+.special {font-weight:bold; font-size: x-large}
+.bad {color: red;}
+.curly {font-family: "Brush Script MT"}
 .toe {margin-left: 1em; font-style: italic;}
 little-hero {color:blue; font-size: smaller; background-color: Turquoise }
