+++ public/docs/ts/latest/guide/user-input.jade
@@ -4,6 +4,8 @@ include ../../../../_includes/_util-fns
   When the user clicks a link, pushes a button, or types on the keyboard
   we want to know about it. These user actions all raise DOM events.
   In this chapter we learn to bind to those events using the Angular Event Binding syntax.
+  
+  [Live Example](/resources/live-examples/user-input/ts/src/plnkr.html).
 
 :marked
   ## Binding to User Input Events
